requirements.txt
transformers
torch
python-dotenv
psycopg2-binary
rich
